+++ src/test/java/com/johnsproject/jpge2/processing/VectorProcessorTest.java
@@ -6,9 +6,9 @@ public class VectorProcessorTest {
 
 	@Test
 	public void multiplyMatrixTest() throws Exception {
-		int[] out = VectorProcessor.generate();
-		int[][] matrix = MatrixProcessor.generate();
-		int[] vector = VectorProcessor.generate(6, 3, 2);
+		long[] out = VectorProcessor.generate();
+		long[][] matrix = MatrixProcessor.generate();
+		long[] vector = VectorProcessor.generate(6, 3, 2);
 		VectorProcessor.multiply(vector, matrix, out);
 		assert(out[0] == 6);
 		assert(out[1] == 3);
@@ -18,14 +18,14 @@ public class VectorProcessorTest {
 	
 	@Test
 	public void magnitudeTest() throws Exception {
-		int[] vector = VectorProcessor.generate(6 << MathProcessor.FP_SHIFT, 3 << MathProcessor.FP_SHIFT, 2 << MathProcessor.FP_SHIFT);
+		long[] vector = VectorProcessor.generate(6 << MathProcessor.FP_SHIFT, 3 << MathProcessor.FP_SHIFT, 2 << MathProcessor.FP_SHIFT);
 		assert(VectorProcessor.magnitude(vector) == 7);
 	}
 	
 	@Test
 	public void normalizeTest() throws Exception {
-		int[] out = VectorProcessor.generate();
-		int[] vector = VectorProcessor.generate(6 << MathProcessor.FP_SHIFT, 3 << MathProcessor.FP_SHIFT, 2 << MathProcessor.FP_SHIFT);
+		long[] out = VectorProcessor.generate();
+		long[] vector = VectorProcessor.generate(6 << MathProcessor.FP_SHIFT, 3 << MathProcessor.FP_SHIFT, 2 << MathProcessor.FP_SHIFT);
 		VectorProcessor.normalize(vector, out);
 		assert(out[0] <= 1024);
 		assert(out[1] <= 1024);
